# [Buildkite] Test Android Staging on `ami-0953ca1d949c23a58` -> `ami-0f976e0b4de96aa28`