--- a/.buildkite/beta-builds.yml
+++ b/.buildkite/beta-builds.yml
@@ -5,7 +5,7 @@
 # only used for beta builds
 
 agents:
-  queue: "android"
+  queue: "android-staging"
 
 steps:
   #################