--- a/.buildkite/schedules/dependency-cache.yml
+++ b/.buildkite/schedules/dependency-cache.yml
@@ -2,7 +2,7 @@
 ---
 
 agents:
-  queue: "android"
+  queue: "android-staging"
 
 steps:
   - label: "dependency cache"