--- a/.buildkite/shared-pipeline-vars
+++ b/.buildkite/shared-pipeline-vars
@@ -3,5 +3,5 @@
  # This file is `source`'d before calling `buildkite-agent pipeline upload`, and can be used
  # to set up some variables that will be interpolated in the `.yml` pipeline before uploading it.
 
-export CI_TOOLKIT="automattic/a8c-ci-toolkit#3.9.1"
+export CI_TOOLKIT="automattic/a8c-ci-toolkit#4aed409fb7535d16c4e08794997fad3773389a9b"
 export TEST_COLLECTOR="test-collector#v1.10.1"