@ParaskP7 2025-01-24
👋 @wzieba same as with [WCAndroid](https://github.com/woocommerce/woocommerce-android/pull/13387#issuecomment-2612631091), please take a look! 🙏