@ParaskP7 2025-01-28
FYI: Just like it was done for [WCAndroid](https://github.com/woocommerce/woocommerce-android/pull/13387#discussion_r1931952101) @wzieba, this is also 05222966e844d128ac469f44ef4e29c03d55ab9f ✅ for this project as well, everything is looking as expected:

- Before: [1563.65 MB](https://buildkite.com/automattic/wordpress-android/builds/20967#019498f1-bb31-4eba-8b4b-9a7bfda1bd43/23409-23421) cache size -> [4122](https://scans.gradle.com/s/x5tvcyzoybbia/performance/network-activity) network requests
- After: [1563.68 MB](https://buildkite.com/automattic/wordpress-android/builds/21026#0194ac93-2fe1-4820-9f83-38fa32d995d7/23407-23419) cache size -> [4108](https://scans.gradle.com/s/22quzciwex6co/performance/network-activity) network requests

PS: Note that although there exist some minor differences, but those are negligible and IMHO could be ignored.